Move helper methods to single file
