+++ x/superfluid/keeper/stake_test.go
@@ -3,13 +3,9 @@ package keeper_test
 import (
 	"time"
 
-	"github.com/cosmos/cosmos-sdk/crypto/keys/secp256k1"
-	"github.com/cosmos/cosmos-sdk/simapp"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/x/staking/teststaking"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 	epochstypes "github.com/osmosis-labs/osmosis/v7/x/epochs/types"
-	lockupkeeper "github.com/osmosis-labs/osmosis/v7/x/lockup/keeper"
 	lockuptypes "github.com/osmosis-labs/osmosis/v7/x/lockup/types"
 	minttypes "github.com/osmosis-labs/osmosis/v7/x/mint/types"
 	"github.com/osmosis-labs/osmosis/v7/x/superfluid/keeper"
@@ -36,43 +32,6 @@ type assetTwap struct {
 type osmoEquivalentMultiplier struct {
 	denom string
 	price sdk.Dec
-}
-
-func (suite *KeeperTestSuite) LockTokens(addr sdk.AccAddress, coins sdk.Coins, duration time.Duration) (lockID uint64) {
-	msgServer := lockupkeeper.NewMsgServerImpl(*suite.app.LockupKeeper)
-	err := simapp.FundAccount(suite.app.BankKeeper, suite.ctx, addr, coins)
-	suite.Require().NoError(err)
-	msgResponse, err := msgServer.LockTokens(sdk.WrapSDKContext(suite.ctx), lockuptypes.NewMsgLockTokens(addr, duration, coins))
-	suite.Require().NoError(err)
-	return msgResponse.ID
-}
-
-func (suite *KeeperTestSuite) SetupValidator(bondStatus stakingtypes.BondStatus) sdk.ValAddress {
-	valPub := secp256k1.GenPrivKey().PubKey()
-	valAddr := sdk.ValAddress(valPub.Address())
-	delAddr := sdk.AccAddress(valAddr)
-	bondDenom := suite.app.StakingKeeper.GetParams(suite.ctx).BondDenom
-	selfBond := sdk.NewCoins(sdk.Coin{Amount: sdk.NewInt(100), Denom: bondDenom})
-
-	simapp.FundAccount(suite.app.BankKeeper, suite.ctx, delAddr, selfBond)
-	sh := teststaking.NewHelper(suite.T(), suite.ctx, *suite.app.StakingKeeper)
-	msg := sh.CreateValidatorMsg(valAddr, valPub, selfBond[0].Amount)
-	sh.Handle(msg, true)
-	val, found := suite.app.StakingKeeper.GetValidator(suite.ctx, valAddr)
-	suite.Require().True(found)
-	val = val.UpdateStatus(bondStatus)
-	suite.app.StakingKeeper.SetValidator(suite.ctx, val)
-
-	return valAddr
-}
-
-func (suite *KeeperTestSuite) SetupValidators(bondStatuses []stakingtypes.BondStatus) []sdk.ValAddress {
-	valAddrs := []sdk.ValAddress{}
-	for _, status := range bondStatuses {
-		valAddr := suite.SetupValidator(status)
-		valAddrs = append(valAddrs, valAddr)
-	}
-	return valAddrs
 }
 
 func (suite *KeeperTestSuite) SetupSuperfluidDelegations(delAddrs []sdk.AccAddress, valAddrs []sdk.ValAddress, superDelegations []superfluidDelegation) ([]types.SuperfluidIntermediaryAccount, []lockuptypes.PeriodLock) {
